propperly refractored
