+++ src/index.css
@@ -2,17 +2,3 @@
   margin: 0;
   padding: 0;
 }
-
-/*
-The logo in the login page
-*/
-
-.login-imag {
-  height: 30vh;
-  width: auto;
-}
-
-.login-icons {
-  margin-right: 0.5vh;
-  margin-top: 0.5vh;
-}
