+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import Login from './components/Login.js';
+import Login from './components/Login/Login.js';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
